package renames and heirarchy changes
put agent and message packages within the weiss.core package, so if a
client doesn't want to use the manager but still wants to use the
backend, they can
