+++ src/weiss/core/agent/Agent.java
@@ -14,12 +14,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package weiss.MetaAgent;
+package weiss.core.agent;
 
 
-import weiss.Message.Message;
-import weiss.Message.SysMessage;
-import weiss.Message.UserMessage;
+import weiss.core.message.Message;
+import weiss.core.message.SysMessage;
+import weiss.core.message.UserMessage;
 
 /**
  *
