+++ src/weiss/core/agent/Monitorable.java
@@ -14,9 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package weiss.MetaAgent;
+package weiss.core.agent;
 
-import Weiss.Manager.NodeMonitor;
+import weiss.manager.NodeMonitor;
 
 /**
  *
