+++ src/weiss/core/agent/WeissBase.java
@@ -14,15 +14,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package weiss.MetaAgent;
+package weiss.core.agent;
 
-import Weiss.Manager.NodeMonitor;
+import weiss.manager.NodeMonitor;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import weiss.Message.Message;
-import weiss.Message.SysMessage;
-import weiss.Message.UserMessage;
+import weiss.core.message.Message;
+import weiss.core.message.SysMessage;
+import weiss.core.message.UserMessage;
 
 /**
  *
