+++ src/weiss/core/message/UserMessage.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package weiss.Message;
+package weiss.core.message;
 
 import java.util.UUID;
